# contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BAHI",
    "description": "The udhaar book both sides can see. All data is synthetic.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/shops/{shop_id}/inbox": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "Get Inbox",
        "description": "C1. Every thread, the latest first, with what is unread.",
        "operationId": "get_inbox_shops__shop_id__inbox_get",
        "parameters": [
          {
            "name": "shop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Shop Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InboxOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/shops/{shop_id}/inbox/read": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Read All",
        "description": "Mark all read.",
        "operationId": "read_all_shops__shop_id__inbox_read_post",
        "parameters": [
          {
            "name": "shop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Shop Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/shops/{shop_id}/customers/{customer_id}/thread": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "Get Shop Thread",
        "description": "C3. His thread, as the shop sees it.",
        "operationId": "get_shop_thread_shops__shop_id__customers__customer_id__thread_get",
        "parameters": [
          {
            "name": "shop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Shop Id"
            }
          },
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Customer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Shop Says",
        "description": "The shopkeeper writes to him. Words only: a message carries no amount.",
        "operationId": "shop_says_shops__shop_id__customers__customer_id__thread_post",
        "parameters": [
          {
            "name": "shop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Shop Id"
            }
          },
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Customer Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SayIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/shops/{shop_id}/customers/{customer_id}/thread/read": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Read Shop Thread",
        "operationId": "read_shop_thread_shops__shop_id__customers__customer_id__thread_read_post",
        "parameters": [
          {
            "name": "shop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Shop Id"
            }
          },
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Customer Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/shops/{shop_id}/customers/{customer_id}/thread/suggest": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Suggest",
        "description": "Replies the munshi would send, in his language. Only suggestions: the\nshopkeeper taps one to send it. None when voice is off.",
        "operationId": "suggest_shops__shop_id__customers__customer_id__thread_suggest_post",
        "parameters": [
          {
            "name": "shop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Shop Id"
            }
          },
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Customer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepliesOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/shops/{shop_id}/entries/{entry_id}/correct": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Correct",
        "description": "C2. The right amount for an entry he says is wrong: a new entry, which he\nconfirms on his own phone. The disputed one is kept, and counts for nothing.",
        "operationId": "correct_shops__shop_id__entries__entry_id__correct_post",
        "parameters": [
          {
            "name": "shop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Shop Id"
            }
          },
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entry Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CorrectIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/shops/{shop_id}/events": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "Get Events",
        "description": "V7. What happened since `after`: for the Soundbox's tones and the screen's\nnews. The first call (no `after`) only says what time it is.",
        "operationId": "get_events_shops__shop_id__events_get",
        "parameters": [
          {
            "name": "shop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Shop Id"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "After"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/people/{person_id}/shops/{shop_id}/thread": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "Get My Thread",
        "description": "C2. His thread with one shop.",
        "operationId": "get_my_thread_people__person_id__shops__shop_id__thread_get",
        "parameters": [
          {
            "name": "person_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Person Id"
            }
          },
          {
            "name": "shop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Shop Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/people/{person_id}/shops/{shop_id}/thread/read": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Read My Thread",
        "operationId": "read_my_thread_people__person_id__shops__shop_id__thread_read_post",
        "parameters": [
          {
            "name": "person_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Person Id"
            }
          },
          {
            "name": "shop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Shop Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/shops/{shop_id}/thread": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Customer Says",
        "description": "He writes to the shop.",
        "operationId": "customer_says_shops__shop_id__thread_post",
        "parameters": [
          {
            "name": "shop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Shop Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerSayIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CorrectIn": {
        "properties": {
          "amount_paise": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Amount Paise"
          }
        },
        "type": "object",
        "required": [
          "amount_paise"
        ],
        "title": "CorrectIn"
      },
...
      "CustomerSayIn": {
        "properties": {
          "person_id": {
            "type": "string",
            "format": "uuid",
            "title": "Person Id"
          },
          "text": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "person_id",
          "text"
        ],
        "title": "CustomerSayIn"
      },
...
      "EntryOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "customer_id": {
            "type": "string",
            "title": "Customer Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "shop_id": {
            "type": "string",
            "title": "Shop Id"
          },
          "shop_name": {
            "type": "string",
            "title": "Shop Name"
          },
          "amount_paise": {
            "type": "integer",
            "title": "Amount Paise"
          },
          "paid_paise": {
            "type": "integer",
            "title": "Paid Paise"
          },
          "status": {
            "type": "string",
            "enum": [
              "recorded",
              "confirmed",
              "disputed",
              "corrected",
              "settled"
            ],
            "title": "Status"
          },
          "recorded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Recorded At"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "spoken_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spoken Text"
          },
          "corrects_entry_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Corrects Entry Id"
          },
          "acknowledged_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acknowledged At"
          },
          "button": {
            "type": "string",
            "title": "Button"
          }
        },
        "type": "object",
        "required": [
          "id",
          "customer_id",
          "display_name",
          "shop_id",
          "shop_name",
          "amount_paise",
          "paid_paise",
          "status",
          "recorded_at",
          "note",
          "spoken_text",
          "corrects_entry_id",
          "acknowledged_at",
          "button"
        ],
        "title": "EntryOut"
      },
      "EventOut": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "scanned",
              "confirmed",
              "disputed",
              "paid",
              "message"
            ],
            "title": "Kind"
          },
          "at": {
            "type": "string",
            "format": "date-time",
            "title": "At"
          },
          "customer_id": {
            "type": "string",
            "title": "Customer Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "amount_paise": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount Paise"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "at",
          "customer_id",
          "display_name",
          "amount_paise"
        ],
        "title": "EventOut",
        "description": "Something the shop should hear about. The Soundbox plays a tone, or says\nan amount; it never says a name. The screen shows who."
      },
      "EventsOut": {
        "properties": {
          "now": {
            "type": "string",
            "format": "date-time",
            "title": "Now"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/EventOut"
            },
            "type": "array",
            "title": "Events"
          }
        },
        "type": "object",
        "required": [
          "now",
          "events"
        ],
        "title": "EventsOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InboxOut": {
        "properties": {
          "today": {
            "type": "string",
            "format": "date",
            "title": "Today"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/InboxRowOut"
            },
            "type": "array",
            "title": "Rows"
          },
          "unread": {
            "type": "integer",
            "title": "Unread"
          }
        },
        "type": "object",
        "required": [
          "today",
          "rows",
          "unread"
        ],
        "title": "InboxOut"
      },
      "InboxRowOut": {
        "properties": {
          "customer_id": {
            "type": "string",
            "title": "Customer Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "tag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag"
          },
          "joined": {
            "type": "string",
            "enum": [
              "linked",
              "invited",
              "name_only"
            ],
            "title": "Joined"
          },
          "author": {
            "type": "string",
            "enum": [
              "shop",
              "customer",
              "bahi"
            ],
            "title": "Author"
          },
          "kind": {
            "type": "string",
            "enum": [
              "text",
              "entry",
              "reminder"
            ],
            "title": "Kind"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "sent_at": {
            "type": "string",
            "format": "date-time",
            "title": "Sent At"
          },
          "entry": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ThreadEntryOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "unread": {
            "type": "integer",
            "title": "Unread"
          },
          "needs_reply": {
            "type": "boolean",
            "title": "Needs Reply"
          },
          "reminder_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reminder At"
          }
        },
        "type": "object",
        "required": [
          "customer_id",
          "display_name",
          "tag",
          "joined",
          "author",
          "kind",
          "body",
          "sent_at",
          "entry",
          "unread",
          "needs_reply",
          "reminder_at"
        ],
        "title": "InboxRowOut"
      },
...
      "MessageOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "author": {
            "type": "string",
            "enum": [
              "shop",
              "customer",
              "bahi"
            ],
            "title": "Author"
          },
          "kind": {
            "type": "string",
            "enum": [
              "text",
              "entry",
              "reminder"
            ],
            "title": "Kind"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "sent_at": {
            "type": "string",
            "format": "date-time",
            "title": "Sent At"
          },
          "entry": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ThreadEntryOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "card": {
            "type": "boolean",
            "title": "Card"
          }
        },
        "type": "object",
        "required": [
          "id",
          "author",
          "kind",
          "body",
          "sent_at",
          "entry",
          "card"
        ],
        "title": "MessageOut"
      },
...
      "RepliesOut": {
        "properties": {
          "replies": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Replies"
          }
        },
        "type": "object",
        "required": [
          "replies"
        ],
        "title": "RepliesOut",
        "description": "Replies the munshi suggests. Only offered: nothing is sent until he taps."
      },
...
      "SayIn": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "SayIn"
      },
...
      "ShopOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "locality": {
            "type": "string",
            "title": "Locality"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "locality"
        ],
        "title": "ShopOut"
      },
...
      "ThreadOut": {
        "properties": {
          "today": {
            "type": "string",
            "format": "date",
            "title": "Today"
          },
          "thread_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thread Id"
          },
          "customer_id": {
            "type": "string",
            "title": "Customer Id"
          },
          "shop": {
            "$ref": "#/components/schemas/ShopOut"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "tag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag"
          },
          "joined": {
            "type": "string",
            "enum": [
              "linked",
              "invited",
              "name_only"
            ],
            "title": "Joined"
          },
          "balance_paise": {
            "type": "integer",
            "title": "Balance Paise"
          },
          "day": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Day"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/MessageOut"
            },
            "type": "array",
            "title": "Messages"
          },
          "reminder_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reminder At"
          }
        },
        "type": "object",
        "required": [
          "today",
          "thread_id",
          "customer_id",
          "shop",
          "display_name",
          "tag",
          "joined",
          "balance_paise",
          "day",
          "messages",
          "reminder_at"
        ],
        "title": "ThreadOut",
        "description": "One shop and one customer, from either side."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}